+++ src/main/presenter/threadPresenter/utils/contentEnricher.ts
