+++ src/main/presenter/threadPresenter/utils/promptBuilder.ts
@@ -16,8 +16,8 @@ import {
   ChatMessageContent
 } from '@shared/presenter'
 import type { MCPToolDefinition } from '@shared/presenter'
-import { ContentEnricher } from '../enrichers/contentEnricher'
-import { buildUserMessageContext, getNormalizedUserMessageText } from '../utils/messageContent'
+import { ContentEnricher } from './contentEnricher'
+import { buildUserMessageContext, getNormalizedUserMessageText } from './messageContent'
 import { generateSearchPrompt } from '../managers/searchManager'
 import { nanoid } from 'nanoid'
